backend(rest_f)/requirements-new.txt
absl-py==1.3.0
appdirs==1.4.3
asgiref==3.5.2
astunparse==1.6.3
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.3
cycler==0.11.0
daphne==4.0.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.2.16
django-cors-headers==3.13.0
django-picklefield==3.1
django-thread==0.0.1
djangorestframework==3.14.0
efficientnet==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.0.12
firebase-admin==6.0.1
flatbuffers==23.5.26
fonttools==4.37.4
gast==0.4.0
gcloud==0.18.3
gdown==4.5.3
gevent==22.10.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.50.0
grpcio-status==1.50.0
h5py==3.7.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
incremental==22.10.0
ipykernel==6.16.0
ipympl==0.9.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
joblib==1.2.0
jsonfield==3.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
jws==0.1.3
keras==2.13.1
Keras-Applications @ git+https://github.com/keras-team/keras-applications.git@06fbeb0f16e1304f239b2296578d1c50b15a983a
Keras-Preprocessing==1.1.2
keras-swa==0.1.7
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.22.4
oauth2client==3.0.0
oauthlib==3.2.2
obspy==1.3.0
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.9.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2019.3
PyWavelets==1.3.0
pywin32==304
pyzmq==24.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
service-identity==21.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.3.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.0.20
wcwidth==0.2.5
websocket==0.2.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.5.2
